phoenix_api/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
six==1.16.0
pytz==2022.6
openai==0.25.0
social-auth-app-django==5.0